Create /fetch serverless handler
`/generate` is an endpoint that you can post a JSON of the country
register and it will return a data file based on just that.

`/fetch` is an endpoint that will go off and fetch the country,
territory, uk registers and the synonym CSV and return a full data
file.