Add IAM policy
This is generated with
https://github.com/dancrumb/generator-serverless-policy.

This policy should only be used by a CI environment for the purposes
of redeploying an already deployed serverless application. It doesn't
have permissions to create buckets or other things that are necessary
for initial serverless bootstrap.

However, I had to perform a number of manual changes:

- Add `cloudformation:ValidateTemplate` as detailed in
https://github.com/dancrumb/generator-serverless-policy/issues/7
- Consolidate s3:*, cloudformation, iam permissions
- Remove unused sqs, kinesis, logs, cloudwatch permissions
- Scope down permissions further so wherever possible they only access
  serverless ARNs